Add files via upload
This Python program is using web scrapping to scrape temperature from timeanddate.com in real-time by using the requests module of Python and showing the temperature and calories required to the user by taking some inputs from the user like its age, weight city location. It can be integrated with flask to make a robust web application. This project is open for contributions.